lua: convert fnl configs to lua
