--- fnl/dotfiles/macros.fnl
@@ -1,11 +1,0 @@
-{:augroup
- (fn [name ...]
-   `(do 
-      (nvim.ex.augroup ,(tostring name))
-      (nvim.ex.autocmd_)
-      ,...
-      (nvim.ex.augroup :END)))
-
- :autocmd
- (fn [...]
-   `(nvim.ex.autocmd ,...))}
