--- fnl/dotfiles/module/plugins/sexp.fnl
@@ -1,4 +1,0 @@
-(module dotfiles.module.plugin.sexp
-  {require {nvim aniseed.nvim}})
-
-(set nvim.g.sexp_filetypes "clojure,scheme,lisp,timl,fennel,janet")
