--- fnl/dotfiles/module/plugins/todo-comments.fnl
@@ -1,5 +1,0 @@
-(module dotfiles.module.plugin.todo-comments
-  {require {nvim aniseed.nvim
-            todo-comments todo-comments}})
-
-(todo-comments.setup)
