--- fnl/dotfiles/module/plugins/tokyonight.fnl
@@ -1,5 +1,0 @@
-(module dotfiles.module.plugin.tokyonight
-  {require {nvim aniseed.nvim}})
-
-(set nvim.g.tokyonight_style :night)
-(nvim.ex.colorscheme "tokyonight")
