--- fnl/dotfiles/module/plugins/trouble.fnl
@@ -1,5 +1,0 @@
-(module dotfiles.module.plugin.trouble
-  {require {nvim aniseed.nvim
-            trouble trouble}})
-
-(trouble.setup)
