+++ lua/atta/main/init.lua
@@ -1,0 +1,11 @@
+local general = require('atta.main.general')
+local keybindings = require('atta.main.keybindings')
+
+local M = {}
+
+function M.setup()
+  general.setup()
+  keybindings.setup()
+end
+
+return M
