+++ lua/atta/plugins/nvim-treesitter.lua
@@ -1,0 +1,18 @@
+local treesitter_configs = require("nvim-treesitter.configs")
+
+local M = {}
+
+function M.setup()
+  treesitter_configs.setup({
+    ensure_installed = "all",
+    ignore_install = { "haskell" },
+    highlight = {
+      enable = true
+    },
+    indent = {
+      enable = true
+    }
+  })
+end
+
+return M
