+++ lua/atta/plugins/sexp.lua
@@ -1,0 +1,9 @@
+local g = vim.g
+
+local M = {}
+
+function M.setup()
+  g.sexp_filetypes = "clojure,scheme,lisp,timl,fennel,janet"
+end
+
+return M
