+++ lua/atta/plugins/todo_comments.lua
@@ -1,0 +1,9 @@
+local todo_comments = require("todo-comments")
+
+local M = {}
+
+function M.setup()
+  todo_comments.setup()
+end
+
+return M
