+++ lua/atta/plugins/tokyonight.lua
@@ -1,0 +1,8 @@
+local M = {}
+
+function M.setup()
+  vim.g.tokyonight_style = "night"
+  vim.cmd([[colorscheme tokyonight]])
+end
+
+return M
