+++ lua/atta/plugins/trouble.lua
@@ -1,0 +1,9 @@
+local trouble = require("trouble")
+
+local M = {}
+
+function M.setup()
+  trouble.setup()
+end
+
+return M
